+++ R/expected-degrees.R
@@ -16,6 +16,26 @@
 #'
 #' @examples
 #'
+#' ##### an undirected blockmodel example
+#'
+#' n <- 1000
+#' pop <- n / 2
+#' a <- .1
+#' b <- .05
+#'
+#' B <- matrix(c(a,b,b,a), nrow = 2)
+#'
+#' b_model <- fastRG::sbm(n = n, k = 2, B = B,  edge_distribution = "bernoulli")
+#'
+#' b_model
+#'
+#' # compare
+#' mean(rowSums(triu(A)))
+#'
+#' pop * a + pop * b  # analytical average degree
+#'
+#' ##### more generic examples
+#'
 #' n <- 10000
 #' k <- 5
 #'
@@ -70,7 +90,7 @@ expected_edges.undirected_factor_model <- function(factor_model, ...) {
   S <- factor_model$S
 
   Cx <- Diagonal(n = ncol(X), x = colSums(X))
-  sum(Cx %*% S %*% Cx) * 2
+  sum(Cx %*% S %*% Cx)
 }
 
 #' @rdname expected_edges
@@ -119,69 +139,7 @@ expected_degrees.undirected_factor_model <- function(factor_model, ...) {
 
 #' @export
 expected_density.undirected_factor_model <- function(factor_model, ...) {
-  expected_edges(factor_model) / as.numeric(factor_model$n)^2
-}
-
-#' @importFrom RSpectra eigs_sym
-#' @export
-RSpectra::eigs_sym
-
-#' @method eigs_sym undirected_factor_model
-#' @export
-eigs_sym.undirected_factor_model <- function(
-  A, k = A$k,
-  which = "LM", sigma = NULL,
-  opts = list(),
-  ...) {
-
-  if (!requireNamespace("RSpectra", quietly = TRUE)) {
-    stop(
-      "Must install `RSpectra` for this functionality.",
-      call. = FALSE
-    )
-  }
-
-  Ax <- function(x, args) as.numeric(args$X %*% (args$SXt %*% x))
-
-  eigs_sym(Ax, k, n = A$n, args = list(X = A$X, SXt = tcrossprod(A$S, A$X)))
-}
-
-#' @method svds undirected_factor_model
-#' @export
-svds.undirected_factor_model <- function(
-    A,
-    k = A$k,
-    nu = k,
-    nv = k,
-    opts = list(),
-    ...) {
-
-  if (!requireNamespace("RSpectra", quietly = TRUE)) {
-    stop(
-      "Must install `RSpectra` for this functionality.",
-      call. = FALSE
-    )
-  }
-
-  Ax <- function(x, args) {
-    as.numeric(args$X %*% (tcrossprod(args$S, args$X) %*% x))
-  }
-
-  Atx <- function(x, args) {
-    as.numeric(tcrossprod(args$X, args$S) %*% crossprod(args$X, x))
-  }
-
-  svds(
-    A = Ax,
-    k = k,
-    nu = nu,
-    nv = nv,
-    opts = opts,
-    ...,
-    Atrans = Atx,
-    dim = c(A$n, A$n),
-    args = list(X = A$X, S = A$S)
-  )
+  expected_edges(factor_model) / as.numeric(choose(factor_model$n, 2))
 }
 
 #' @export
@@ -201,49 +159,4 @@ expected_density.directed_factor_model <- function(factor_model, ...) {
   d <- factor_model$d
 
   expected_edges(factor_model) / (as.numeric(n) * as.numeric(d))
-}
-
-#' @importFrom RSpectra svds
-#' @export
-RSpectra::svds
-
-
-
-
-#' @method svds directed_factor_model
-#' @export
-svds.directed_factor_model <- function(
-  A,
-  k = min(A$k1, A$k2),
-  nu = k,
-  nv = k,
-  opts = list(),
-  ...) {
-
-  if (!requireNamespace("RSpectra", quietly = TRUE)) {
-    stop(
-      "Must install `RSpectra` for this functionality.",
-      call. = FALSE
-    )
-  }
-
-  Ax <- function(x, args) {
-    as.numeric(args$X %*% (tcrossprod(args$S, args$Y) %*% x))
-  }
-
-  Atx <- function(x, args) {
-    as.numeric(tcrossprod(args$Y, args$S) %*% crossprod(args$X, x))
-  }
-
-  svds(
-    A = Ax,
-    k = k,
-    nu = nu,
-    nv = nv,
-    opts = opts,
-    ...,
-    Atrans = Atx,
-    dim = c(A$n, A$d),
-    args = list(X = A$X, S = A$S, Y = A$Y)
-  )
 }