+++ NEWS.md
@@ -5,6 +5,7 @@
 * Allow rank 1 undirected stochastic block models
 * Fix bug where isolated nodes where dropped from sampled tidygraphs (#23)
 * Allow users to force model identification in DC-SBMs with `force_identifiability = TRUE`, and in overlapping SBMs with `force_pure = TRUE`, which are now the default.
+* Improve population expected degree/density computations (#19)
 
 # fastRG 0.3.0